refactor: split CLI commands into focused files
