+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Changed
 - Aligned local development, documentation, and the pinned Docker builder on Go 1.26.5; added canonical Make targets, tool-neutral contributor guidance, and reproducible CI tool versions.
+- Split the Cobra CLI implementation into focused command files without changing command behavior.
 - Bumped google.golang.org/grpc from 1.82.0 to 1.82.1 (semver-patch) for bug fixes and security updates
 - Bumped modernc.org/sqlite from 1.54.0 to 1.55.0 (semver-minor) for upstream bug fixes
 - Bumped google.golang.org/grpc from 1.82.1 to 1.83.0 (semver-minor) for upstream bug fixes
